+++ src/fhir_scripts/deploy.py
@@ -6,6 +6,7 @@ from . import log
 from .helper import confirm
 from .models.config import Config, DeployConfig
 from .tools import gcloud
+from .types import Url
 
 TARGET_BASE_DIR = "ig/fhir"
 
@@ -14,6 +15,14 @@ def setup_parser(parser: ArgumentParser, *args, **kwarsg):
     parser.add_argument("environment", help="Name of the environment")
     parser.add_argument(
         "-y", "--yes", action="store_true", help="Confirm all prompts with 'yes'"
+    )
+    parser.add_argument(
+        "--promote-from",
+        type=str,
+        help="Promote from another environment insread of the local build directory",
+    )
+    parser.add_argument(
+        "--dry-run", action="store_true", help="Only pretend to deploy files"
     )
 
     group = parser.add_mutually_exclusive_group()
@@ -27,10 +36,10 @@ def deploy(
     config: Config,
     environment: str,
     ig_registry: bool = False,
-    all: bool = False,
     only_ig: bool = False,
-    only_history: bool = False,
     yes: bool = False,
+    dry_run: bool = False,
+    promote_from: str | None = None,
     *args,
     **kwargs,
 ) -> bool:
@@ -39,151 +48,212 @@ def deploy(
         raise Exception("deploy configuration missing")
 
     if ig_registry:
-        # Build target URL
-        target = _target_path(deploy_config, environment)
-        _deploy_ig_registry(target, confirm_yes=yes)
+        deploy_ig_registry(
+            deploy_config,
+            environment,
+            promote_from_env=promote_from,
+            dry_run=dry_run,
+            confirm_yes=yes,
+        )
 
     else:
-        # Build target URL
-        target = _target_path(deploy_config, environment, needs_project=True)
-
         if only_ig:
-            _deploy_ig(target, confirm_yes=yes)
+            deploy_ig(
+                deploy_config,
+                environment,
+                promote_from_env=promote_from,
+                dry_run=dry_run,
+                confirm_yes=yes,
+            )
 
         else:
-            _deploy_ig(target, confirm_yes=yes)
-            _deploy_history(target, confirm_yes=yes)
-            _deploy_package_list(target, confirm_yes=yes)
+            deploy_ig(
+                deploy_config,
+                environment,
+                promote_from_env=promote_from,
+                dry_run=dry_run,
+                confirm_yes=yes,
+            )
+            deploy_ig_meta(
+                deploy_config,
+                environment,
+                promote_from_env=promote_from,
+                dry_run=dry_run,
+                confirm_yes=yes,
+            )
 
     return True
 
 
-def _deploy_ig_registry(target, confirm_yes: bool = False):
-    reg_dir = Path("./")
+def deploy_ig_registry(
+    deploy_cfg: DeployConfig,
+    target_env: str,
+    promote_from_env: str | None = None,
+    confirm_yes: bool = False,
+    dry_run: bool = False,
+    *args,
+    **kwargs,
+):
+    target_path = get_storage_path(deploy_cfg, target_env)
 
-    log.info(f"Deploy IG registry files to {target}")
+    if promote_from_env:
+        source_path = get_storage_path(deploy_cfg, promote_from_env)
+        log.info("Promote IG registry files {} -> {}".format(source_path, target_path))
+
+    else:
+        source_path = Path("./")
+        log.info("Deploy IG registry files -> {}".format(target_path))
+
     confirm("Continue?", "Aborted by user", confirm_yes=confirm_yes, default=True)
 
     deploy_files = ["index.html", "package-feed.xml"]
+    files = []
     for file in deploy_files:
+        source_file = source_path / file
+        target_file = target_path / file
 
-        file = reg_dir / file
-        if not file.exists():
-            raise Exception(f"Source '{file.absolute()}' does not exist")
+        files.append((source_file, target_file))
 
-        target_file = target + "/" + file.name
-        gcloud.copy(source=file, target=target_file, force=True)
+        if dry_run:
+            log.info("Would have copied {} -> {}".format(source_file, target_file))
+
+        else:
+            gcloud.copy(source=source_file, target=target_file, force=True)
 
     log.succ("Deployed IG registry files")
+    return files
 
 
-def _target_path(
-    deploy_cfg: DeployConfig, env_name: str, needs_project: bool = False
-) -> str:
-    env = deploy_cfg.env.get(env_name)
+def deploy_ig(
+    deploy_cfg: DeployConfig,
+    target_env: str,
+    promote_from_env: str | None = None,
+    confirm_yes: bool = False,
+    dry_run: bool = False,
+    *args,
+    **kwargs,
+):
 
-    if env is None:
-        raise Exception(f"Environment '{env_name}' not defined in config")
+    if promote_from_env:
+        # Get 'project' and 'version' from publication request
+        project, version = project_version_from_pub_req()
 
-    path_parts: list[str]
-    if deploy_cfg.path is not None:
-        path_parts = [env, deploy_cfg.path]
+        source_path = get_storage_path(deploy_cfg, promote_from_env) / project / version
+        target_path = get_storage_path(deploy_cfg, target_env) / project / version
 
-    elif project := _project_name(needs_project):
-        path_parts = [env, TARGET_BASE_DIR, project]
+        log.info("Promote built IG {} -> {}".format(source_path, target_path))
 
     else:
-        if needs_project:
-            raise Exception(
-                "Path needs to include a project name, none found in built IG or config"
-            )
+        # Get 'project' and 'version' from built implementation guide
+        project, version = project_version_from_imp_guide()
 
-        path_parts = [env, TARGET_BASE_DIR]
+        source_path = Path("./output")
+        target_path = get_storage_path(deploy_cfg, target_env) / project / version
 
-    return "gs://" + "/".join([p.strip("/") for p in path_parts])
+        log.info("Deploy built IG -> {}".format(target_path))
+
+    # Copy IG
+    confirm("Continue?", "Aborted by user", confirm_yes=confirm_yes, default=True)
+
+    if dry_run:
+        log.info("Would have copied {} -> {}".format(source_path, target_path))
+
+    else:
+        gcloud.copy(source=source_path, target=target_path, force=confirm_yes)
+
+    log.succ("Deployed IG")
+
+    return source_path, target_path
 
 
-def _deploy_ig(target: str, confirm_yes: bool = False):
+def deploy_ig_meta(
+    deploy_cfg: DeployConfig,
+    target_env: str,
+    promote_from_env: str | None = None,
+    dry_run: bool = False,
+    *args,
+    **kwargs,
+):
+
+    if promote_from_env:
+        # Get 'project' and 'version' from poblication request
+        project, _ = project_version_from_pub_req()
+
+        source_path = get_storage_path(deploy_cfg, promote_from_env) / project
+        target_path = get_storage_path(deploy_cfg, target_env) / project
+
+        log.info("Promote IG meta data {} -> {}".format(source_path, target_path))
+
+    else:
+        # Get 'project' and 'version' from built implementation guide
+        project, _ = project_version_from_imp_guide()
+
+        # Check possible source paths
+        # First the current path
+        if ((path := Path("./publish")) / "index.html").exists():
+            source_path = path
+
+        elif ((path := path / project) / "index.html").exists():
+            source_path = path
+
+        else:
+            log.warn("Meta files not found, skipping")
+            return []
+
+        target_path = get_storage_path(deploy_cfg, target_env) / project
+
+        log.info("Deploy IG meta data -> {}".format(target_path))
+
+    # Copy meta files
+    deploy_files = ["index.html", "package-list.json"]
+    files = []
+    for file in deploy_files:
+        source_file = source_path / file
+        target_file = target_path / file
+
+        files.append((source_file, target_file))
+
+        if dry_run:
+            log.info("Would have copied {} -> {}".format(source_file, target_file))
+
+        else:
+            gcloud.copy(source=source_file, target=target_file, force=True)
+
+    log.succ("Deployed IG meta files")
+    return files
+
+
+def get_storage_path(deploy_cfg: DeployConfig, env_name: str) -> Url:
+    storage = deploy_cfg.env.get(env_name)
+
+    if storage is None:
+        raise Exception(f"Environment '{env_name}' not defined in config")
+
+    storage = Url(storage) if storage.startswith("gs://") else Url("gs://" + storage)
+
+    return storage / (deploy_cfg.path or TARGET_BASE_DIR)
+
+
+def project_version_from_pub_req() -> tuple[str, str]:
+    pub_req = Path(".") / "publication-request.json"
+
+    if not pub_req.exists():
+        raise Exception("Publication request missing")
+
+    req = json.loads(pub_req.read_text(encoding="utf-8"))
+    _, project, version = req["path"].rsplit("/", 2)
+
+    return project, version
+
+
+def project_version_from_imp_guide() -> tuple[str, str]:
     output_dir = Path("./output")
-
-    # Get built version
     igs = list(output_dir.glob("ImplementationGuide*.json"))
     if len(igs) != 1:
         raise Exception("Built IG not found")
 
     ig = json.loads(igs[0].read_text(encoding="utf-8"))
-    version = ig["version"]
-
-    # Copy IG
-    target_versioned = f"{target}/{version}"
-    log.info(f"Deploy built IG to {target_versioned}")
-    confirm("Continue?", "Aborted by user", confirm_yes=confirm_yes, default=True)
-    gcloud.copy(source=output_dir, target=target_versioned, force=confirm_yes)
-
-    log.succ("Deployed IG")
-
-
-def _deploy_history(target, *args, **kwargs):
-    history_file_name = "index.html"
-    publish_dir = Path("./publish")
-
-    if (file := publish_dir / history_file_name).exists():
-        history_file = file
-
-    # Old file location
-    elif (
-        file := publish_dir / _project_name(needs_project=True) / history_file_name
-    ).exists():
-        history_file = file
-
-    else:
-        log.warn("History file not found, skipping")
-        return
-
-    # Copy history
-    target_history = target + "/" + history_file_name
-    log.info(f"Deploy history file to {target_history}")
-    gcloud.copy(source=history_file, target=target_history, force=True)
-
-    log.succ("Deployed history file")
-
-
-def _project_name(needs_project: bool = True) -> str | None:
-    """
-    Extract the project name from a IG Canonical URL
-    """
-    output_dir = Path("./output")
-
-    # Get project name
-    igs = list(output_dir.glob("ImplementationGuide*.json"))
-    if len(igs) != 1:
-        if needs_project:
-            raise Exception("Built IG not found")
-
-        else:
-            return None
-
-    ig = json.loads(igs[0].read_text(encoding="utf-8"))
-    return ig["url"].rsplit("/", 3)[1]
-
-
-def _deploy_package_list(target, *args, **kwargs):
-    package_list_file_name = "package-list.json"
-    publish_dir = Path("./publish")
-
-    if (file := publish_dir / package_list_file_name).exists():
-        package_list_file = file
-
-    else:
-        log.warn("History file not found, skipping")
-        return
-
-    # Copy history
-    target_file = target + "/" + package_list_file_name
-    log.info(f"Deploy package list file to {target_file}")
-    gcloud.copy(source=package_list_file, target=target_file, force=True)
-
-    log.succ("Deployed package list file")
+    return ig["url"].rsplit("/", 3)[1], ig["version"]
 
 
 __doc__ = "Deploy IG"